+++ c-expr-dsl/test/Test/CExpr/Parse.hs
@@ -4,9 +4,9 @@ module Test.CExpr.Parse (
 
 import Test.Tasty
 
-import Test.CExpr.Parse.Golden    qualified as Golden
-import Test.CExpr.Parse.Macro     qualified as Macro
-import Test.CExpr.Parse.Type      qualified as Type
+import Test.CExpr.Parse.Golden qualified as Golden
+import Test.CExpr.Parse.Macro qualified as Macro
+import Test.CExpr.Parse.Type qualified as Type
 
 tests :: TestTree
 tests = testGroup "parse" [